pyproject.toml
[project]
name = "coordinate-transformation-api"
description = "Coordinate Transformation API of the Nederlandse Samenwerking Geodetische Infrastructuur (NSGI)"
keywords = ["NSGI", "pyproj", "fastapi", "coordinates", "geo"]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
]
readme = { file = "./README.md", content-type = "text/markdown" }
dependencies = [
    "fastapi[all] == 0.113.0",
    "uvicorn == 0.30.6",
    "pyyaml == 6.0.2",
    "geojson_pydantic == 1.1.1",
    "pyproj == 3.6.1",
    "pydantic-settings == 2.4.0",
    "email-validator == 2.2.0",
    "geodense == 2.0.0",
]
requires-python = ">=3.11.4"
dynamic = ["version"]


[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project.scripts]
ct-api = "coordinate_transformation_api.main:main"

[tool.hatch.version]
source = "vcs"


[tool.ruff]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "build",
]

line-length = 120

# Allow unused variables when underscore-prefixed.
target-version = "py39"

[tool.ruff.lint]

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

select = [
    "E",
    "F",
    "B",
    "ISC",
    "PL",
    "RUF",
    "PT",
    "N",
    "I",
    "C90",
    "W",
    "UP",
    "S",
    "ARG",
    "SIM",
    "Q",
    "ANN",
]
# Avoid enforcing line-length violations (`E501`) - fixed by Black
ignore = ["E501"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = [
    "E",
    "F",
    "B",
    "ISC",
    "PL",
    "RUF",
    "PT",
    "N",
    "I",
    "C90",
    "W",
    "UP",
    "S",
    "ARG",
    "SIM",
    "Q",
]
unfixable = []

# disable rules for tests
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101", # asserts allowed
    "ANN",  # no type annotation required
]
"src/coordinate_transformation_api/cityjson/models.py" = [
    "N815", # non-snakecase variables allowed
]
"src/coordinate_transformation_api/util.py" = ["ANN"]

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10


[tool.uv]
dev-dependencies = [
    "pre-commit>=4.0.0",
    "ruff>=0.6.4",
    "yamllint>=1.35.1",
    "coverage == 7.6.1",
    "mypy == 1.11.2",
    "pytest == 8.3.2",
    "ruff == 0.6.4",
    "types-ujson == 5.10.0.20240515",
    "types-PyYAML == 6.0.12.20240808",
    "pytest-asyncio == 0.24.0",
]


[tool.mypy]
plugins = "pydantic.mypy"

warn_return_any = true
warn_unused_configs = true


[[tool.mypy.overrides]]
module = ["mypy-shapely.*", "mypy-geodense.*"]
ignore_missing_imports = true
